+++ src/britive/workload.py
@@ -1,3 +1,6 @@
+from . import exceptions
+from typing import Union
+
 
 class Workload:
     def __init__(self, britive):
@@ -35,6 +38,27 @@ class Workload:
 
             return self.britive.get(f'{self.base_url}/{workload_identity_provider_id}')
 
+        def _build_attributes_map_list(self, attributes_map: dict):
+            # first get list of existing custom identity attributes and build some helpers
+            existing_attrs = [attr for attr in self.britive.identity_attributes.list() if not attr['builtIn']]
+            existing_attr_ids = [attr['id'] for attr in existing_attrs]
+            attrs_by_name = {attr['name']: attr['id'] for attr in existing_attrs}
+
+            # for each attributeMap key/value provided ensure we convert to ID and build the list
+            attrs_list = []
+            for idp_attr, custom_identity_attribute in attributes_map.items():
+                if custom_identity_attribute not in existing_attr_ids:
+                    custom_identity_attribute = attrs_by_name.get(custom_identity_attribute, None)
+                if not custom_identity_attribute:
+                    raise ValueError(f'custom identity attribute name {custom_identity_attribute} not found.')
+                attrs_list.append(
+                    {
+                        'idpAttr': idp_attr,
+                        'userAttr': custom_identity_attribute
+                    }
+                )
+            return attrs_list
+
         def create(self, **kwargs) -> dict:
             """
             Create a workload identity provider.
@@ -46,6 +70,8 @@ class Workload:
             Generally, the caller should opt to use the `create_aws` or `create_oidc` methods instead of calling
             this method directly.
 
+            The field `attributesMap` is in format `{'idp_attr_value': 'custom_identity_attribute_name_or_id', ...}`.
+
             :param kwargs: A list of keyword arguments which will be provided directly to the API backend without
                 any further inspection. Valid fields are `name`, `description`, `idpType`, `attributesMap`, and
                 `validationWindow`. For the AWS provider an additional field `maxDuration` is valid. For the OIDC
@@ -53,16 +79,20 @@ class Workload:
             :returns: Details of the newly created workload identity provider.
             """
 
+            if 'attributesMap' in kwargs.keys():
+                kwargs['attributesMap'] = self._build_attributes_map_list(attributes_map=kwargs['attributesMap'])
+
             return self.britive.post(self.base_url, json=kwargs)
 
-        def create_aws(self, name: str, attributes_map: dict = None, description: str = None,
+        def create_aws(self, name: str, attributes_map: dict, description: str = None,
                        validation_window: int = 30, max_duration: int = 5) -> dict:
             """
             Create an AWS workload identity provider.
 
             :param name: Name of the AWS workload identity provider.
             :param attributes_map: The mapping of the available AWS workload token fields to custom identity attributes
-                which are associated with service identities. Defaults to None.
+                which are associated with service identities. Defaults to None.Can provide either the name of the
+                custom identity attribute or the ID.
             :param description: Optional description of the AWS workload identity provider.
             :param validation_window: The number of seconds allowed to validate the AWS workload token from its
                 "issued time". Defaults to 30 seconds.
@@ -86,12 +116,13 @@ class Workload:
         def create_oidc(self, name: str, issuer_url: str, attributes_map: dict = None, description: str = None,
                         validation_window: int = 30, allowed_audiences: list = None) -> dict:
             """
-            Create an AWS workload identity provider.
+            Create an OIDC workload identity provider.
 
             :param name: Name of the OIDC workload identity provider.
             :param issuer_url: The issuer url for the OIDC provider.
             :param attributes_map: The mapping of the available OIDC workload token fields to custom identity attributes
-                which are associated with service identities. Defaults to None.
+                which are associated with service identities. Defaults to None. Can provide either the name of the
+                custom identity attribute or the ID.
             :param description: Optional description of the AWS workload identity provider.
             :param validation_window: The number of seconds allowed to validate the AWS workload token from its
                 "issued time". Defaults to 30 seconds.
@@ -116,15 +147,111 @@ class Workload:
 
         def update(self, workload_identity_provider_id: int, **kwargs) -> dict:
             """
-            Updates details of the specified workload identity provider.
+            Updates a workload identity provider.
 
-            :param workload_identity_provider_id: The ID of the workload identity provider.
-            :param kwargs: Any attributes of the workload identity provider that should be updated. Reference the
-                `create` method for details on what attributes are available.
+            This method accepts a list of kwargs which match the required and optional fields for creating a
+            workload identity provider. It mostly exists to support the update of new workload identity provider
+            types before this SDK can be updated to support the new type natively.
+
+            Generally, the caller should opt to use the `update_aws` or `update_oidc` methods instead of calling
+            this method directly.
+
+            The field `attributesMap` is in format `{'idp_attr_value': 'custom_identity_attribute_name_or_id', ...}`.
+
+            :param workload_identity_provider_id: The ID of the workload identity provider to update.
+            :param kwargs: A list of keyword arguments which will be provided directly to the API backend without
+                any further inspection. Valid fields are `name`, `description`, `idpType`, `attributesMap`, and
+                `validationWindow`. For the AWS provider an additional field `maxDuration` is valid. For the OIDC
+                provider additional fields `issuerUrl` and `allowedAudiences` are valid.
             :returns: Details of the updated workload identity provider.
             """
+
             kwargs['id'] = workload_identity_provider_id
-            return self.britive.put(self.base_url, json=kwargs)
+
+            if 'attributesMap' in kwargs.keys():
+                kwargs['attributesMap'] = self._build_attributes_map_list(attributes_map=kwargs['attributesMap'])
+
+            # since this is a PUT call and not a PATCH call we need to get the existing idp configuration
+            # and merge in the things that have changed
+
+            existing = self.get(workload_identity_provider_id=workload_identity_provider_id)
+            return self.britive.put(self.base_url, json={**existing, **kwargs})
+
+        def update_aws(self, workload_identity_provider_id: int, name: str = None, attributes_map: dict = None,
+                       description: str = None, validation_window: int = None, max_duration: int = None) -> dict:
+            """
+            Update an AWS workload identity provider.
+
+            All fields except `workload_identity_provider_id` are optional.
+
+            :param workload_identity_provider_id: The ID of the workload identity provider to update.
+            :param name: Name of the AWS workload identity provider.
+            :param attributes_map: The mapping of the available AWS workload token fields to custom identity attributes
+                which are associated with service identities.Can provide either the name of the custom identity
+                attribute or the ID.
+            :param description: Description of the AWS workload identity provider.
+            :param validation_window: The number of seconds allowed to validate the AWS workload token from its
+                "issued time".
+            :param max_duration: The max number of hours (whole numbers only) for which the AWS workload token is valid.
+            :returns: Details of the updated AWS workload identity provider.
+            """
+
+            params = {
+                'idpType': 'AWS'
+            }
+
+            if name:
+                params['name'] = name
+            if description:
+                params['description'] = description
+            if validation_window:
+                params['validationWindow'] = validation_window
+            if max_duration:
+                params['maxDuration'] = max_duration
+            if attributes_map:
+                params['attributesMap'] = attributes_map
+
+            return self.update(workload_identity_provider_id=workload_identity_provider_id, **params)
+
+        def update_oidc(self, workload_identity_provider_id: int, name: str = None, issuer_url: str = None,
+                        attributes_map: dict = None, description: str = None, validation_window: int = None,
+                        allowed_audiences: list = None) -> dict:
+            """
+            Update an OIDC workload identity provider.
+
+            All fields except `workload_identity_provider_id` are optional.
+
+            :param workload_identity_provider_id: The ID of the workload identity provider to update.
+            :param name: Name of the OIDC workload identity provider.
+            :param issuer_url: The issuer url for the OIDC provider.
+            :param attributes_map: The mapping of the available OIDC workload token fields to custom identity attributes
+                which are associated with service identities. Can provide either the name of the custom identity
+                attribute or the ID.
+            :param description: Description of the AWS workload identity provider.
+            :param validation_window: The number of seconds allowed to validate the AWS workload token from its
+                "issued time".
+            :param allowed_audiences: The list of allowed audience values as strings.
+            :returns: Details of the update OIDC workload identity provider.
+            """
+
+            params = {
+                'idpType': 'OIDC'
+            }
+
+            if name:
+                params['name'] = name
+            if description:
+                params['description'] = description
+            if validation_window:
+                params['validationWindow'] = validation_window
+            if issuer_url:
+                params['issuerUrl'] = issuer_url
+            if attributes_map:
+                params['attributesMap'] = attributes_map
+            if allowed_audiences:
+                params['allowedAudiences'] = allowed_audiences
+
+            return self.update(workload_identity_provider_id=workload_identity_provider_id, **params)
 
         def delete(self, workload_identity_provider_id) -> None:
             """
@@ -133,13 +260,16 @@ class Workload:
             :param workload_identity_provider_id: The ID of the workload identity provider.
             :returns: None.
             """
-            self.britive.delete(f'{self.base_url}/{workload_identity_provider_id}')
-            return None
+            return self.britive.delete(f'{self.base_url}/{workload_identity_provider_id}')
 
         def generate_attribute_map(self, idp_attribute_name: str, custom_identity_attribute_name: str = None,
                                    custom_identity_attribute_id: str = None) -> dict:
             """
-            Generates a dictionary that can be appended to a list used for the attributes_map.
+            Generates a dictionary that can be appended to a list used for the `attributesMap`.
+
+            This method would mostly be used when invoking the `create` or `update` methods directly instead of
+            using the type specific (`create_aws`, `create_oidc`, `update_aws`, `update_oidc`) methods which provide
+            a more pythonic way to capture the attribute mappings.
 
             :param idp_attribute_name: The name of the workload identity provider attribute to map. This will always
                 be a string as it is controlled by the identity provider.
@@ -180,18 +310,22 @@ class Workload:
             self.britive = workload.britive
             self.base_url: str = workload.base_url + '/users/{id}/identity-provider'  # will .format(id=...) later
 
-        def get(self, service_identity_id: str) -> dict:
+        def get(self, service_identity_id: str) -> Union[dict, None]:
             """
             Returns details about the workload identity provider associated with the specified service identity.
 
             :param service_identity_id: The ID of the service identity.
             :returns: Details about the workload identity provider associated with the specified service identity.
+                Returns None if there is no workload identity provider associated with the specified service identity.
             """
 
-            return self.britive.get(self.base_url.format(id=service_identity_id))
+            try:
+                return self.britive.get(self.base_url.format(id=service_identity_id))
+            except exceptions.NotFound:
+                return None
 
-        def assign(self, service_identity_id: str, idp_id: str, token_duration: int = 300,
-                   federated_attributes_ids: dict = None, federated_attributes_names: dict = None) -> dict:
+        def assign(self, service_identity_id: str, idp_id: str, federated_attributes: dict,
+                   token_duration: int = 300) -> dict:
             """
             Associates an OIDC provider with the specified Service Identity.
 
@@ -200,13 +334,9 @@ class Workload:
             :param token_duration: Duration in seconds (from now) before a token provided by the client expires.
                 This will be evaluated alongside the OIDC JWT expiration field and the earlier of the two values
                 will govern when the token expires. This field is optional and defaults to 300 seconds.
-            :param federated_attributes_ids: An attribute map where keys are the custom attribute ids and values are
-                strings or list of strings (for multivalued attributes) which map back to the mapped token claims.
-                This will be merged with `federated_attributes_names` and if there are duplicates this will win.
-            :param federated_attributes_names: An attribute map where keys are the custom attribute names and values are
-                strings or list of strings (for multivalued attributes) which map back to the mapped token claims.
-                The names will be auto-converted to ids and merged with `federated_attributes_ids`. If there are
-                duplicates `federated_attributes_ids` will win.
+            :param federated_attributes: An attribute map where keys are the custom attribute ids or names
+                and values are strings or list of strings (for multivalued attributes) which map back to the mapped
+                token claims.
             :returns: Details of the newly assigned workload identity provider.
             """
 
@@ -217,18 +347,16 @@ class Workload:
 
             response = self.britive.post(self.base_url.format(id=service_identity_id), json=params)
 
-            if federated_attributes_ids or federated_attributes_names:
-                try:
-                    self.britive.service_identities.set_custom_identity_attributes(
-                        service_identity_id=service_identity_id,
-                        custom_attributes_ids=federated_attributes_ids,
-                        custom_attributes_names=federated_attributes_names
-                    )
-                except Exception as e:
-                    # need to remove the assignment as something went wrong with the attributes mapping
-                    self.unassign_idp(service_identity_id=service_identity_id)
-                    # and re-raise the error
-                    raise e
+            try:
+                self.britive.service_identities.set_custom_identity_attributes(
+                    service_identity_id=service_identity_id,
+                    custom_attributes=federated_attributes,
+                )
+            except Exception as e:
+                # need to remove the assignment as something went wrong with the attributes mapping
+                self.unassign(service_identity_id=service_identity_id)
+                # and re-raise the error
+                raise e
 
             return response
 
@@ -247,7 +375,7 @@ class Workload:
             )
             self.britive.service_identities.remove_custom_identity_attributes(
                 service_identity_id=service_identity_id,
-                custom_attributes_ids=existing_attributes
+                custom_attributes=existing_attributes
             )
             return self.britive.delete(self.base_url.format(id=service_identity_id))