+++ tests/test_080-scans.py
@@ -2,12 +2,14 @@ import time
 from .cache import *  # will also import some globals like `britive`
 
 
+@pytest.mark.skipif(scan_skip, reason=scan_skip_message)
 def test_scan(cached_scan):
     assert isinstance(cached_scan, dict)
     assert 'taskId' in cached_scan.keys()
 
 
 # warning - this will take a while if a scan was just initiated! grab and coffee and come back later
+@pytest.mark.skipif(scan_skip, reason=scan_skip_message)
 def test_status(cached_scan):
     while True:
         status = britive.scans.status(task_id=cached_scan['taskId'])
@@ -19,12 +21,14 @@ def test_status(cached_scan):
     assert status['status'] == 'Success'
 
 
+@pytest.mark.skipif(scan_skip, reason=scan_skip_message)
 def test_history(cached_application):
     response = britive.scans.history(application_id=cached_application['appContainerId'])
     assert isinstance(response, list)
     assert len(response) > 0
 
 
+@pytest.mark.skipif(scan_skip, reason=scan_skip_message)
 def test_diff_accounts(cached_application, cached_environment):
     response = britive.scans.diff(
         resource='accounts',
@@ -35,6 +39,7 @@ def test_diff_accounts(cached_application, cached_environment):
     assert len(response) > 0
 
 
+@pytest.mark.skipif(scan_skip, reason=scan_skip_message)
 def test_diff_groups(cached_application, cached_environment):
     response = britive.scans.diff(
         resource='groups',
@@ -45,6 +50,7 @@ def test_diff_groups(cached_application, cached_environment):
     assert len(response) > 0
 
 
+@pytest.mark.skipif(scan_skip, reason=scan_skip_message)
 def test_diff_permissions(cached_application, cached_environment):
     response = britive.scans.diff(
         resource='permissions',